GBL unit fixes (#402)
* GBL unit fixes

* Updated EUTelTripletGBLUtility::FindTriplets

* Sanitizing units

* Experimental SLC6 Dockerfile

* Docker fix

* Werror=catch-value fixes, #404

* Updated test cases for changes in EUTelGeo stuff

* More catch-value errors fixed

* Updated Dockerfiles

* Testing ctest values

* Ctest SLC test cases updated

+++ processors/src/EUTelProcessorSparseClustering.cc
@@ -549,7 +549,7 @@ void EUTelProcessorSparseClustering::bookHistos() {
     // minY, maxX and maxY.
     int minX, minY, maxX, maxY;
     minX = minY = maxX = maxY = 0;
-    float sizeX, sizeY, sizeZ;
+    double sizeX, sizeY, sizeZ;
     sizeX = sizeY = sizeZ = 0;
 
     geoDescr->getPixelIndexRange(minX, maxX, minY, maxY);
